Extract ScopedMMap from cuttlefish/common/libs:fs
Bug: b/543998370
