+++ base/cvd/cuttlefish/host/libs/audio_connector/server.cpp
@@ -30,6 +30,7 @@
 #include "absl/log/check.h"
 #include "absl/log/log.h"
 
+#include "cuttlefish/common/libs/fs/scoped_mmap.h"
 #include "cuttlefish/common/libs/fs/shared_fd.h"
 #include "cuttlefish/host/libs/audio_connector/buffers.h"
 #include "cuttlefish/host/libs/audio_connector/commands.h"
