+++ base/cvd/cuttlefish/host/libs/screen_connector/BUILD.bazel
@@ -45,6 +45,7 @@ cf_cc_library(
         ":video_frame_buffer",
         "//cuttlefish/common/libs/concurrency",
         "//cuttlefish/common/libs/fs",
+        "//cuttlefish/common/libs/fs:scoped_mmap",
         "//cuttlefish/common/libs/utils:contains",
         "//cuttlefish/common/libs/utils:size_utils",
         "//cuttlefish/host/libs/config:cuttlefish_config",
